requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0

# CLIP (OpenAI)
# Install separately:
#   pip install git+https://github.com/openai/CLIP.git
ftfy
regex
tqdm

# Simulation (optional — install only what you use)
# gymnasium>=0.29.0
# gym-robotics>=1.3.0          # FrankaKitchen, FetchReach
# minigrid>=2.3.0              # MiniGrid envs

# Visualization
matplotlib>=3.7.0
tensorboard>=2.14.0
